Touchpad emulator
Takes an evemu recording as input and the path to a real touchpad device. Then
it creates a uinput device based on the evemu recording and maps any incoming
data from the real touchpad device to the matching position in the evemu
device.

Can be used to debug pointer motion or other issues on different touchpads
when only one touchpad is available.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>